% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_all_tx.R
\name{align_all_tx}
\alias{align_all_tx}
\title{Returns alignments between an RNA sequence and all supplied DNA sequences}
\usage{
align_all_tx(rna, dna, code_list, ka_list, species)
}
\arguments{
\item{rna}{RNA sequences of transcripts of interest along with metadata. List.}

\item{dna}{DNA sequences of interest along with metadata. List.}

\item{code_list}{list of substitution matrices returned by \code{get_codes()}}

\item{ka_list}{list of Karlin-Altschul parameters (list) returned by \code{get_ka_params()}.}

\item{species}{"human"/"hs" or "mouse"/"mm".}
}
\description{
This function aligns two sequences using a substitution matrix learned by exprectation-maximisation from triplex-sequencing data, and returns a \code{data.frame} with results including Karlin-Altschul statistics.
}
\keyword{align}
\keyword{dna}
\keyword{rna}
\keyword{triplex}
