% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_per_code.R
\name{align_per_code}
\alias{align_per_code}
\title{Returns alignments for each substitution matrix in a list}
\usage{
align_per_code(rna_seq, dna_seq, code_list, ka_list, species)
}
\arguments{
\item{rna_seq}{RNA sequence of interest}

\item{dna_seq}{DNA sequence of interest}

\item{code_list}{list of substitution matrices returned by \code{get_codes()}}

\item{ka_list}{list of Karlin-Altschul parameters (list) returned by \code{get_ka_params()}.}

\item{species}{"human"/"hs" or "mouse"/"mm".}
}
\description{
This function aligns two sequences using a substitution matrix learned by exprectation-maximisation from triplex-sequencing data, and returns a \code{data.frame} with results including Karlin-Altschul statistics.
}
\keyword{align}
\keyword{dna}
\keyword{rna}
\keyword{triplex}
