% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_score.R
\name{bit_score}
\alias{bit_score}
\title{Return bit-score of local alignment}
\usage{
bit_score(score, lambda, k)
}
\arguments{
\item{score}{raw score from local alignment.}

\item{lambda}{Karlin-Altschul parameter.}

\item{k}{Karlin-Altschul parameter.}
}
\description{
This function will return either the human or mouse UCSC transcript databases
}
\examples{
bit_score(score = 50, lambda = 0.35, k = 0.2)
}
\keyword{Altschul}
\keyword{Karlin}
\keyword{alignment}
\keyword{score}
