% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_value.R
\name{e_value}
\alias{e_value}
\title{Return \emph{E} value of local alignment bit-score}
\usage{
e_value(bit_score, m, n, N)
}
\arguments{
\item{bit_score}{bit_score from local alignment, returned by \code{bit_score()}.}

\item{m}{Karlin-Altschul parameter, DNA sequence length.}

\item{n}{Karlin-Altschul parameter, RNA sequence length.}

\item{N}{Karlin-Altschul parameter, RNA database size (i.e. total transcriptome length).}
}
\description{
This function will return the \emph{E} value of a local alignment, calculated using Karlin-Altschul statistics
}
\examples{
e_value(bit_score = 27.5, m = 3000, n = 750, N = 8323091082)
}
\keyword{Altschul}
\keyword{Karlin}
\keyword{alignment}
\keyword{score}
