% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bs.R
\name{get_bs}
\alias{get_bs}
\title{Return species appropriate BSgenome object}
\usage{
get_bs(species)
}
\arguments{
\item{species}{either "human" or "hs" for human, or "mouse" or "mm" for mouse.}
}
\description{
This function will return either the human or mouse UCSC BSGenome
}
\examples{
get_bs("human") # returns human bs
get_bs("hs") # returns human bs
get_bs("mouse") # returns mouse bs
get_bs("mm") # returns mouse bs
}
\keyword{BSgenome}
\keyword{genome}
\keyword{species}
