% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_codes}
\alias{get_codes}
\title{Returns log-odds substitution matrices learned from triplex-sequencing data}
\usage{
get_codes()
}
\description{
Returns the substitution matrices which were learned by expectation-maximisation from triplexRNA-seq and triplexDNA-seq data.
}
\examples{
get_codes()
}
\keyword{code}
\keyword{log-odds}
\keyword{matrix}
