% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ka_params.R
\name{get_ka_params}
\alias{get_ka_params}
\title{Returns Karlin-Altschul parameters for TriplexAligner substitution matrices}
\usage{
get_ka_params()
}
\description{
Returns the Karlin-Altschul parameters for substitution matrices which were learned by expectation-maximisation from triplexRNA-seq and triplexDNA-seq data.
}
\examples{
get_ka_params()
}
\keyword{Altschul}
\keyword{Karlin}
\keyword{code}
\keyword{log-odds}
\keyword{matrix}
\keyword{parameters}
