% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org.R
\name{get_org}
\alias{get_org}
\title{Return species appropriate org.db}
\usage{
get_org(species)
}
\arguments{
\item{species}{either "human" or "hs" for human, or "mouse" or "mm" for mouse.}
}
\description{
This function will return either the human or mouse org.db object
}
\examples{
get_org("human") # returns human org
get_org("hs") # returns human org
get_org("mouse") # returns mouse org
get_org("mm") # returns mouse org
}
\keyword{org}
\keyword{species}
\keyword{transcripts}
