% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transcriptome_length.R
\name{get_transcriptome_length}
\alias{get_transcriptome_length}
\title{Returns the total transcriptome length of the appropriate species}
\usage{
get_transcriptome_length(species)
}
\arguments{
\item{species}{raw score from local alignment.}
}
\description{
This function will return the total length of either the human or mouse transcriptome, as annotated by the UCSC TxDB packages.
}
\examples{
# Get length of annotated human transcriptome
get_transcriptome_length(species = "human")
# Get length of annotated mouse transcriptome
get_transcriptome_length(species = "mouse")
}
\keyword{length}
\keyword{transcriptome}
