% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txdb.R
\name{get_txdb}
\alias{get_txdb}
\title{Return species appropriate transcript database}
\usage{
get_txdb(species)
}
\arguments{
\item{species}{either "human" or "hs" for human, or "mouse" or "mm" for mouse.}
}
\description{
This function will return either the human or mouse UCSC transcript databases
}
\examples{
get_txdb("human") # returns human txdb
get_txdb("hs") # returns human txdb
get_txdb("mouse") # returns mouse txdb
get_txdb("mm") # returns mouse txdb
}
\keyword{species}
\keyword{transcripts}
\keyword{txdb}
