% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_alignment.R
\name{parse_alignment}
\alias{parse_alignment}
\title{Parses an alignment output from \code{align} to a \code{data.frame} with key metrics.}
\usage{
parse_alignment(alignment)
}
\arguments{
\item{alignment}{output of \code{align}.}
}
\description{
This function parses the output of \code{align} to a \code{data.frame} which contains key results.
}
\examples{
\dontrun{
parse_alignment(alignment)
}
}
\keyword{align}
\keyword{data.frame}
\keyword{result}
