% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_DNA_from_bed.R
\name{parse_DNA_from_bed}
\alias{parse_DNA_from_bed}
\title{Returns DNA sequences as a DNAstringset for supplied genomic regions}
\usage{
parse_DNA_from_bed(bed_file, species, up = 0, down = 0)
}
\arguments{
\item{bed_file}{character vector of gene symbols e.g. c("GAPDH", "MALAT1") for human or c("Gapdh", "Malat1") for mouse.}

\item{species}{either "human" or "hs" for human, or "mouse" or "mm" for mouse.}

\item{up}{numeric, how many base pairs upstream of the transcription start site to consider. Default 0.}

\item{down}{numeric, how many base pairs downstream of the transcription start site to consider. Default 0.}
}
\description{
Uses TxDb, org.db and BSgenome objects from the species of interest to retrieve promoter sequences of the supplied genes, flanked by a user-defined number of base pairs.
}
\examples{
# return DNA sequences of regions in bed file
parse_proms_from_symbols(bed = foo.bed, species = "human")
# return DNA sequences of regions in bed file exctended by 500 base pairs in each direction
parse_proms_from_symbols(bed = foo.bed, species = "human", up = 500, down = 500)
}
\keyword{gene}
\keyword{promoter}
\keyword{sequence}
\keyword{symbol}
