% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_proms_from_symbols.R
\name{parse_proms_from_symbols}
\alias{parse_proms_from_symbols}
\title{Returns promoter sequences as a DNAstring for supplied gene symbols}
\usage{
parse_proms_from_symbols(symbols, species, up = 2500, down = 500)
}
\arguments{
\item{symbols}{character vector of gene symbols e.g. c("GAPDH", "MALAT1") for human or c("Gapdh", "Malat1") for mouse.}

\item{species}{either "human" or "hs" for human, or "mouse" or "mm" for mouse.}

\item{up}{numeric, how many base pairs upstream of the transcription start site to consider.}

\item{down}{numeric, how many base pairs downstream of the transcription start site to consider.}
}
\description{
Uses TxDb, org.db and BSgenome objects from the species of interest to retrieve promoter sequences of the supplied genes, flanked by a user-defined number of base pairs.
}
\examples{
# return promoter sequences (2500 bp upstream and 500 bp downstream of TSS) of several human genes
parse_proms_from_symbols(symbols = c("GAPDH","CDH5","NANOG","NOXO1"), species = "human", up = 2500, down = 500)
# return promoter sequences (1000 bp upstream and 1000 bp downstream of TSS) of several mouse genes
parse_proms_from_symbols(symbols = c("Gapdh","Cdh5","Nanog","Noxo1"), species = "mouse", up = 1000, down = 1000)
}
\keyword{gene}
\keyword{promoter}
\keyword{sequence}
\keyword{symbol}
