% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tx_from_fasta.R
\name{parse_tx_from_fasta}
\alias{parse_tx_from_fasta}
\title{Returns transcript sequences as a DNAstringset from supplied fasta filepath}
\usage{
parse_tx_from_fasta(fasta)
}
\arguments{
\item{fasta}{Filepath to fasta file of RNAs of interest.}
}
\description{
Uses \code{Biostrings} to parse transcript sequences from the supplied fasta file.
}
\examples{
# return a DNAstringset object containing sequences contained within foo.fa
parse_tx_from_fasta(fasta = foo.fa)
}
\keyword{fasta}
\keyword{gene}
\keyword{sequence}
\keyword{symbol}
\keyword{transcript}
