% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tx_from_symbol.R
\name{parse_tx_from_symbol}
\alias{parse_tx_from_symbol}
\title{Returns transcript sequence as a DNAstring from supplied gene symbol}
\usage{
parse_tx_from_symbol(symbol, species)
}
\arguments{
\item{symbol}{Gene symbol e.g. "GAPDH", "MALAT1" for human or "Gapdh", "Malat1" for mouse.}

\item{species}{either "human" or "hs" for human, or "mouse" or "mm" for mouse.}
}
\description{
Uses TxDb, org.db and BSgenome objects from the species of interest to retrieve transcript sequences of the supplied gene.
}
\examples{
# return all human NEAT1 transcript sequences
parse_tx_from_symbol(symbol = "NEAT1", species = "human")
# return all mouse NEAT1 transcript sequences
parse_tx_from_symbol(symbol = "Neat1", species = "mouse")
}
\keyword{gene}
\keyword{sequence}
\keyword{symbol}
\keyword{transcript}
