% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriplexAligner.R
\name{TriplexAligner}
\alias{TriplexAligner}
\title{Predicts RNA:DNA:DNA triple helix formation between RNAs and DNA regions of interest}
\usage{
TriplexAligner(
  rna_input,
  dna_input,
  rna_format,
  dna_format,
  code_list = get_codes(),
  ka_list = get_ka_params(),
  species,
  up = 2500,
  down = 500
)
}
\arguments{
\item{rna_input}{Input RNA. See \code{rna_format} for options.}

\item{dna_input}{Input DNA. See \code{dna_format} for options.}

\item{rna_format}{One of "symbol" or "fasta".}

\item{dna_format}{One of "symbol", "fasta" or "bed".}

\item{code_list}{list of substitution matrices returned by \code{get_codes()}}

\item{ka_list}{list of Karlin-Altschul parameters (list) returned by \code{get_ka_params()}.}

\item{species}{"human"/"hs" or "mouse"/"mm".}

\item{up}{How many base pairs upstream of TSS to consider as promoter region. Default: 2500.}

\item{down}{How many base pairs downstream of TSS to consider as promoter region. Default: 500.}
}
\description{
This function implements RNA:DNA:DNA triple helix mapping codes learned by expectation-maximisation from triplex-sequencing data as substitution matrices in local alignment to predict sites of RNA-DNA interaction.
}
\examples{
\dontrun{
# Predict triplex formation between MALAT1 and the promoter of GAPDH
TriplexAligner(rna_input = 'MALAT1', dna_input = 'GAPDH', rna_format = 'symbol', dna_format = 'symbol', species = 'hs')
# Predict triplex formation between MALAT1 and DNA regions in a bed file (e.g. ATAC-sequencing peaks)
TriplexAligner(rna_input = 'MALAT1', dna_input = 'foo.bed', rna_format = 'symbol', dna_format = 'bed', species = 'hs')
# Predict triplex formation between MALAT1 & NEAT1 and DNA regions in a fasta file
TriplexAligner(rna_input = c('MALAT1', 'NEAT1'), dna_input = 'foo.fa', rna_format = 'symbol', dna_format = 'bed', species = 'hs')
}
}
\keyword{align}
\keyword{dna}
\keyword{rna}
\keyword{triplex}
